refactor: 💡 rename file to types
